Fix headings in article
